admission: move ioLoadListener into its own file
Pure code movement change. Trying to see what components in this package
could be reasoned about independently to readers less familiar with this
package. Some of these may benefit from being pulled into sub-packages,
but for now a separate file is sufficient.

Release note: None
