+++ pkg/util/admission/store_token_estimation.go
@@ -399,3 +399,10 @@ func (e *storePerWorkTokenEstimator) getModelsAtAdmittedDone() (
 		e.atDoneL0IngestTokensLinearModel.smoothedLinearModel,
 		e.atDoneIngestTokensLinearModel.smoothedLinearModel
 }
+
+func max(i, j int64) int64 {
+	if i < j {
+		return j
+	}
+	return i
+}
